Split React components into different files
